Express.js Routes & Controllers
Routes & Controllers
- api (opentdb)
- index (compression bundle.js & /* route redirect)
- middleware (cors, json, distribution)

applied utils folder
multiple choice random generation and direct deposit
