+++ server/db/connection.js
@@ -1,0 +1,4 @@
+// MongoDB
+var MongoClient = require("mongodb").MongoClient;
+let url = "mongodb://localhost:27017/";
+var mongodb = require("mongodb");
