+++ server/util/shuffle.js
@@ -1,0 +1,16 @@
+shuffle = (array) => {
+  var currentIndex = array.length;
+  var temporaryValue, randomIndex;
+
+  while (0 !== currentIndex) {
+    randomIndex = Math.floor(Math.random() * currentIndex);
+    currentIndex -= 1;
+
+    temporaryValue = array[currentIndex];
+    array[currentIndex] = array[randomIndex];
+    array[randomIndex] = temporaryValue;
+  }
+  return array;
+};
+
+module.exports = { shuffle };
